+++ eNMS/controller.py
@@ -905,6 +905,8 @@ class Controller:
         for model in models:
             path = vs.path / "files" / folder / kwargs["name"] / f"{model}.yaml"
             if not path.exists():
+                if kwargs.get("service_import") and model == "service":
+                    raise Exception("Invalid archive provided in service import.")
                 continue
             with open(path, "r") as migration_file:
                 instances = yaml.load(migration_file)
@@ -995,15 +997,19 @@ class Controller:
         file.save(str(filepath))
         with open_tar(filepath) as tar_file:
             tar_file.extractall(path=vs.path / "files" / "services")
+            folder_name = tar_file.getmembers()[0].name
             status = self.migration_import(
                 folder="services",
-                name=filepath.stem,
+                name=folder_name,
                 import_export_types=["service", "workflow_edge"],
                 service_import=True,
                 skip_pool_update=True,
                 skip_model_update=True,
                 update_pools=True,
             )
+        rmtree(vs.path / "files" / "services" / folder_name, ignore_errors=True)
+        if "Partial import" in status:
+            raise Exception(status)
         return status
 
     def import_topology(self, **kwargs):
@@ -1056,6 +1062,8 @@ class Controller:
 
     @staticmethod
     def run(service, **kwargs):
+        if "path" not in kwargs:
+            kwargs["path"] = str(service)
         keys = list(vs.model_properties["run"]) + list(vs.relationships["run"])
         run_kwargs = {key: kwargs.pop(key) for key in keys if kwargs.get(key)}
         for property in ("name", "labels"):
@@ -1071,6 +1079,7 @@ class Controller:
         restart_run = db.fetch("run", allow_none=True, runtime=restart_runtime)
         if service.type == "workflow" and service.superworkflow and not restart_run:
             run_kwargs["placeholder"] = run_kwargs["start_service"] = service.id
+            run_kwargs["path"] = str(service.superworkflow.id)
             service = service.superworkflow
             initial_payload.update(service.initial_payload)
         else: