+++ eNMS/environment.py
@@ -13,7 +13,7 @@ from logging.config import dictConfig
 from logging import getLogger, info
 from os import getenv, getpid
 from passlib.hash import argon2
-from psutil import cpu_percent, virtual_memory
+from psutil import Process
 from redis import Redis
 from redis.exceptions import ConnectionError, TimeoutError
 from requests import Session as RequestSession
@@ -71,15 +71,14 @@ class Environment:
         class Handler(FileSystemEventHandler):
             def on_any_event(self, event):
                 filetype = "folder" if event.is_directory else "file"
-                if event.event_type in ("deleted", "modified", "moved"):
-                    file = db.fetch(filetype, path=event.src_path, allow_none=True)
-                    if file and event.event_type == "moved":
-                        file.update(path=event.dest_path, move_file=False)
-                elif event.event_type == "created":
+                file = db.fetch(filetype, path=event.src_path, allow_none=True)
+                if event.event_type == "moved" and file:
+                    file.update(path=event.dest_path, move_file=False)
+                elif event.event_type in ("created", "modified"):
                     file = db.factory(filetype, path=event.src_path)
+                elif event.event_type == "deleted" and file:
+                    db.delete_instance(file)
                 else:
-                    return
-                if not file:
                     return
                 file.status = event.event_type.capitalize()
                 log = f"File {event.src_path} {event.event_type} (watchdog)."
@@ -102,6 +101,12 @@ class Environment:
         default_method = vs.settings["authentication"]["default"]
         user_method = getattr(user, "authentication", default_method)
         method = kwargs.get("authentication_method", user_method)
+        if (
+            vs.settings["authentication"]["force_authentication_method"]
+            and user
+            and user.authentication != method
+        ):
+            return False
         if method not in vs.settings["authentication"]["methods"]:
             return False
         elif method == "database":
@@ -224,8 +229,7 @@ class Environment:
         if not self.redis_queue:
             return
         key = f"workers/{getpid()}"
-        self.redis("set", f"{key}/info/memory", f"{virtual_memory().percent}%")
-        self.redis("set", f"{key}/info/cpu", f"{cpu_percent()}%")
+        self.redis("set", f"{key}/info/memory", f"{Process().memory_percent()}%")
         self.redis(mode, f"{key}/jobs/{job}", 1)
 
     def log(self, severity, content, user=None, change_log=True, logger="root"):