+++ eNMS/models/services/data_validation/scrapli.py
@@ -53,7 +53,9 @@ class ScrapliCommandsForm(ScrapliForm):
     form_type = HiddenField(default="scrapli_service")
     commands = StringField(substitution=True, widget=TextArea(), render_kw={"rows": 5})
     jinja2_template = BooleanField(
-        "Interpret Commands as Jinja2 Template", default=False
+        "Interpret Commands as Jinja2 Template",
+        default=False,
+        help="common/commands_jinja",
     )
     results_as_list = BooleanField("Results As List", default=False)
     groups = {