+++ eNMS/static/js/base.js
@@ -313,7 +313,7 @@ export function openPanel({
     id: panelId,
     border: "2px solid #2A3F52",
     theme: panelThemes[name] || "light filledlight",
-    headerLogo: "../static/img/logo.png",
+    headerLogo: "/static/img/logo.png",
     contentOverflow: "hidden scroll",
     contentSize: size || {
       width: () => window.innerWidth * 0.5,
@@ -335,7 +335,7 @@ export function openPanel({
     kwargs.content = content;
   } else {
     kwargs.contentAjax = {
-      url: url || `../${name}_form`,
+      url: url || `/${name}_form`,
       done: function (_, panel) {
         panel.content.innerHTML = this.responseText;
         preprocessForm(panel, id, type, duplicate);
@@ -489,7 +489,7 @@ export function preprocessForm(panel, id, type, duplicate) {
         name: `help-${propertyName}`,
         title: propertyName.replace("_", " "),
         size: "600px auto",
-        url: helpUrl.charAt(0) === "/" ? `..${helpUrl}` : `../help/${helpUrl}`,
+        url: `/help/${helpUrl}`,
         callback: function (helpPanel) {
           helpPanel.querySelectorAll(".help-snippet").forEach((el) => {
             const editor = CodeMirror.fromTextArea(el, {
@@ -973,6 +973,7 @@ export function notify(...args) {
 function showAllAlerts() {
   openPanel({
     name: "alerts_table",
+    title: "Alerts",
     callback: () => {
       $("#alerts-table")
         // eslint-disable-next-line new-cap