+++ eNMS/static/js/visualization.js
@@ -3,7 +3,6 @@ global
 action: false
 L: false
 page: false
-theme: false
 visualization: false
 */