+++ eNMS/variables.py
@@ -136,6 +136,7 @@ class VariableStore:
     def set_template_context(self):
         self.template_context = {
             "application_path": str(self.path),
+            "automation": self.automation,
             "configuration_properties": self.configuration_properties,
             "form_properties": self.form_properties,
             "names": self.property_names,