+++ docs/source/base/release_notes.md
@@ -63,9 +63,9 @@ Version 4.3.0
 - Add User landing page to decide which page to display after logging in (editable in profile).
   Default landing page is configurable from settings.json > authentication > landing_page.
 - Add mechanism to show a single device status in workflow builder UI (logs filtering + service display)
-- Add mechanism to search for a string accross all services of a workflow in the workflow builder, and
+- Add mechanism to search for a string across all services of a workflow in the workflow builder, and
   accross all nodes in the network builder.
-- Fix vertical aligment in all tables (cell content was not centered on y axis because of buttons height in
+- Fix vertical alignment in all tables (cell content was not centered on y axis because of buttons height in
   the last column).
 - Add export service button in Workflow Builder.
 - New Files Management System:
@@ -89,12 +89,23 @@ Version 4.3.0
 - Add new "mandatory" keyword in custom properties to make the field required to submit the form.
 - Add new "show_password_in_profile" keyword in settings > authentication to configure whether the user
   profile lets users change their own password (if `false`, the password field is not shown)
+- Add new "force_authentication_method" to force users to log in with the authentication method saved in
+  the database (e.g first authentication method used)
 - Add new 'Man Minutes' feature to compute time saved per workflow
   * Only for top-level workflows
   * Man Minutes can be defined per device or for the whole workflow
   * Per Device is only allowed if the workflow run method is DxD or SxS with workflow targets
   * The workflow must be a success (or per device success) to be counted in the total man minutes
   * Man Minutes can be made mandatory via 'mandatory_man_minutes' key in automation.json > workflow
+- Remove unused parent and parent_device relationship on the Run class.
+- Import Services:
+  * The timeout for the Import_services endpoint is configurable in "automation.json" under
+    the "service_import" > "timeout" property. Logging on timeout is also improved.
+  * The "stem" of the imported file (e.g., service.tgz) does not have to exactly match the
+    directory in the .tgz file (i.e., "serviceA_v1.tgz" with "serviceA/service.yaml" is supported).
+- The napalm ping service separated the `ping_timeout` from the napalm `timeout`.
+- Add new settings "max_content_length" in settings.json > "app" (Flask parameter)
+- Add new timeout setting for file import in settings.json > "files"
 
 Migration
 - check "username" and "server" variables in workflow aren't in conflict with existing workflows.
@@ -103,6 +114,8 @@ Migration
 - "download_file" endpoint -> "download" (add support for downloading folders)
 - the "driver" property must be updated for all netmiko, napalm and scrapli via the migration script
 - update services to use server IP and address from global variables and not from settings.
+- the napalm_ping_service added a `ping_timeout` property. If desired, set both
+  values to be at least the defaults (2 for `ping_timeout`, 10 for napalm's `timeout`)
 
 To be tested:
 - bulk deletion and bulk removal (from dashboard and other tables too)